fix: don't cross package boundaries for templates
and copy the templates into the expected place in the bundle
