--- packages/cdktf-cli/src/bin/cmds/helper/init-templates.ts
@@ -1,21 +1,0 @@
-// Copyright (c) HashiCorp, Inc
-// SPDX-License-Identifier: MPL-2.0
-import * as fs from "fs-extra";
-import * as path from "path";
-import { projectRootPath } from "./utilities";
-
-export const templatesDir = path.join(
-  projectRootPath(),
-  "..",
-  "@cdktf",
-  "cli-core",
-  "templates"
-);
-
-const availableTemplates = fs
-  .readdirSync(templatesDir)
-  .filter((x) => !x.startsWith("."));
-export const templates: string[] = [];
-for (const template of availableTemplates) {
-  templates.push(template);
-}
